+++ R/fsiStarter.R
@@ -101,7 +101,8 @@ fsiStarter <- function(x, db, grpBy_quo = NULL, scaleBy_quo = NULL, polys = NULL
                                   db$COND$SITECLCD, db$COND$RESERVCD)
   # Tree type
   db$TREE$typeD <- treeTypeDomain(treeType, db$TREE$STATUSCD,
-                                  db$TREE$DIA, db$TREE$TREECLCD)
+                                  db$TREE$DIA, db$TREE$TREECLCD,
+                                  db$TREE$STANDING_DEAD_CD)
   
   # Spatial boundary (determine which of the plots fall within the polygons
   # supplied in polys)